Remove `utils` module.
