+++ crates/revm/src/db/ethersdb.rs
@@ -2,10 +2,42 @@ use std::sync::Arc;
 
 use ethers_core::types::{BlockId, H160 as eH160, H256, U64 as eU64};
 use ethers_providers::Middleware;
+use tokio::runtime::{Builder, Handle, RuntimeFlavor};
 
+use crate::primitives::{AccountInfo, Address, Bytecode, B256, KECCAK_EMPTY, U256};
 use crate::{Database, DatabaseRef};
-use crate::primitives::{AccountInfo, Address, B256, Bytecode, KECCAK_EMPTY, U256};
-use crate::utils::tokio_block_on;
+
+/// internal utility function to call tokio feature and wait for output
+#[inline]
+fn block_on<F>(f: F) -> F::Output
+where
+    F: core::future::Future + Send,
+    F::Output: Send,
+{
+    match Handle::try_current() {
+        Ok(handle) => match handle.runtime_flavor() {
+            // This essentially equals to tokio::task::spawn_blocking because tokio doesn't
+            // allow current_thread runtime to block_in_place
+            RuntimeFlavor::CurrentThread => std::thread::scope(move |s| {
+                s.spawn(move || {
+                    Builder::new_current_thread()
+                        .enable_all()
+                        .build()
+                        .unwrap()
+                        .block_on(f)
+                })
+                .join()
+                .unwrap()
+            }),
+            _ => tokio::task::block_in_place(move || handle.block_on(f)),
+        },
+        Err(_) => Builder::new_current_thread()
+            .enable_all()
+            .build()
+            .unwrap()
+            .block_on(f),
+    }
+}
 
 #[derive(Debug, Clone)]
 pub struct EthersDB<M: Middleware> {
@@ -19,9 +51,7 @@ impl<M: Middleware> EthersDB<M> {
         let block_number: Option<BlockId> = if block_number.is_some() {
             block_number
         } else {
-            Some(BlockId::from(
-                tokio_block_on(client.get_block_number()).ok()?,
-            ))
+            Some(BlockId::from(block_on(client.get_block_number()).ok()?))
         };
 
         Some(Self {
@@ -40,7 +70,7 @@ impl<M: Middleware> DatabaseRef for EthersDB<M> {
     type Error = M::Error;
 
     fn basic_ref(&self, address: Address) -> Result<Option<AccountInfo>, Self::Error> {
-        let add = eH160::from(address.0.0);
+        let add = eH160::from(address.0 .0);
 
         let f = async {
             let nonce = self.client.get_transaction_count(add, self.block_number);
@@ -48,7 +78,7 @@ impl<M: Middleware> DatabaseRef for EthersDB<M> {
             let code = self.client.get_code(add, self.block_number);
             tokio::join!(nonce, balance, code)
         };
-        let (nonce, balance, code) = tokio_block_on(f);
+        let (nonce, balance, code) = block_on(f);
 
         let balance = U256::from_limbs(balance?.0);
         let nonce = nonce?.as_u64();
@@ -63,7 +93,7 @@ impl<M: Middleware> DatabaseRef for EthersDB<M> {
     }
 
     fn storage_ref(&self, address: Address, index: U256) -> Result<U256, Self::Error> {
-        let add = eH160::from(address.0.0);
+        let add = eH160::from(address.0 .0);
         let index = H256::from(index.to_be_bytes());
         let f = async {
             let storage = self
@@ -72,7 +102,7 @@ impl<M: Middleware> DatabaseRef for EthersDB<M> {
                 .await?;
             Ok(U256::from_be_bytes(storage.to_fixed_bytes()))
         };
-        tokio_block_on(f)
+        block_on(f)
     }
 
     fn block_hash_ref(&self, number: U256) -> Result<B256, Self::Error> {
@@ -84,7 +114,7 @@ impl<M: Middleware> DatabaseRef for EthersDB<M> {
         let number = eU64::from(u64::try_from(number).unwrap());
         let f = async { self.client.get_block(BlockId::from(number)).await };
         // If number is given, the block is supposed to be finalized so unwrap is safe too.
-        Ok(B256::new(tokio_block_on(f)?.unwrap().hash.unwrap().0))
+        Ok(B256::new(block_on(f)?.unwrap().hash.unwrap().0))
     }
 }
 
@@ -115,23 +145,22 @@ impl<M: Middleware> Database for EthersDB<M> {
 // Run tests with `cargo test -- --nocapture` to see print statements
 #[cfg(test)]
 mod tests {
-    use ethers_providers::{Http, Provider};
-
     use super::*;
+    use ethers_providers::{Http, Provider};
 
     //#[test]
     fn _can_get_basic() {
         let client = Provider::<Http>::try_from(
             "https://mainnet.infura.io/v3/c60b0bb42f8a4c6481ecd229eddaca27",
         )
-            .unwrap();
+        .unwrap();
         let client = Arc::new(client);
 
         let ethersdb = EthersDB::new(
             Arc::clone(&client), // public infura mainnet
             Some(BlockId::from(16148323)),
         )
-            .unwrap();
+        .unwrap();
 
         // ETH/USDT pair on Uniswap V2
         let address = "0x0d4a11d5EEaaC28EC3F61d100daF4d40471f1852"
